Feature/abilities (#32)
* Provoke abilities

* Stat adjustments

* Remove CON bonus to HP calculations. Add a base of 40 HP to account for the NWN levels applied to the player.

* Fix XP calculation on character sheet

* Update module

* New tlk entries for xp gains and level ups

* XP gains and ability acquisition

* Hook up Defense Bonus

* Clean up usings

* Killer Instinct, Resist Poison, Parry, HP Bonus, Shield Mastery

* Subtle blow

* Critical Bonus

* Accuracy bonus, ether attack bonus, clear mind.

* EP Bonus

* Fix default appearance scale to 1.0f

* Display skills in the skills tab if the player has the lore ability equipped.

* Item equipment restrictions

* Weapon lore abilities

* Natural HP and EP regen. Auto Regen and Auto Refresh abilities

* Fix resonance cost on Auto Refresh

* Barrage and Martial Arts

* Recycle ability

* Check both current job level and filter job level when determining if ability can be equipped.

* Power Attack 1 + 2

* Restoration

* Add icons for Aggressor

* Change status effects over to StatGroups

* Move status effects to progression. Include status calcs in the stat methods.

* TLK update

* Mark feats as target self-only

* Aggressor

* Rest: Only display healing if HP was actually recovered.

* Fix issue where AI would error out if friendly died nearby.

* AI: Don't target dead friendlies

* Fix TLK IDs on spells/feats

* Ward abilities

* Fix tlks on effect icons

* Status effect VFX, Rampart. Recover EP with /rez too

* Add lore abilities to job acquisition list.

* Add support for toggle abilities

* Change Aggressor to Toggle ability

* Raise

* Mighty strikes

* Fix feat check for Lore abilities

* Poison status effect

* Antidote

* Move status effects to combat project

* Protection 1-3

* Divine seal

* Regen

* Benediction

* Etherwarp

* Haste and Slow effects

* Haste ability

* Ether Blast

* Erase ability

* Snarl

* Third Eye

* Remove Tame and give Call Beast to Beastmaster.

* Spell service

* Add way to override stat used.

* Add damage type entries. Change animation for casting. Add Radiant Blast 1 and 2.

* Tsu

* Revive

* Hundred Fists

* Fix Revive definition

* Add Flash

* Warcry

* Chi

* Treasure Hunter and Creditfinder

* Perfect Dodge

* Stasis Field

* Back Attack 1-4

* Sneak Attack

* Flee

* Hide

* Steal

* Eagle Eye Shot

* Sharpshot

* Make Sharpshot stackable from other sources.

* Manafont

* Escape

* Elemental Seal

* Fix Elemental Seal costs and recast group

* Neurotoxin

* Zephyr Shroud

* Ether Wall

* Aurion Infusion

* Convert

* Mindstream

* Queued ability calculations

* Electric Fist 1-3

* Striking Cobra and add poison resistance check

* Change striking cobra to wind

* Remove Venom Stab from Nightstalker

* Hook up telegraphs to ability system

* Null Lance

* Synaptic Shock

* Cripple

* Cryo Venom

* Freezing Shot

* Ice Volley

* Choking Shot

* Shadowstrike

* Flame

* Shockwave Surge

* Shocking Circle & Paralyze effect

* Clamp attributes to 3-127

* Cleanup shadowstrike

* Drown

* Graviton Field

* Organize buffs/debuffs

* Add more debuffs to Erase

* Abyssal Veil

* Cerebral Spike

* Rupture

* Neural Stasis

* Neural Cascade and Pacified

* Add elemental seal to damage checks on spells.

* Crescent Moon

* Mercy Strike

* Move all AOEs over to telegraph system

* Quickness

* Feral Howl

* Beast summoning foundation & Reward

* Call Beast & Broth item

* Remove unused item properties

* Item properties & npc stat loading

* Beast stat application

* Beast speed

* Add builder option to retarget to another creature for beast abilities. Assault

* Ether Link

* Familiar

* Sic

+++ XM.Shared.Progression/UI/CharacterSheet/CharacterSheetViewModel.cs
@@ -14,7 +14,6 @@ using XM.Shared.Core.Entity;
 using XM.Shared.Core.EventManagement;
 using XM.Shared.Core.Localization;
 using XM.UI;
-using XM.UI.Event;
 using Action = System.Action;
 
 namespace XM.Progression.UI.CharacterSheet
@@ -398,7 +397,7 @@ namespace XM.Progression.UI.CharacterSheet
 
                 jobIcons.Add(definition.IconResref);
 
-                var ratio = dbPlayerJob.JobXP[type] / XP[dbPlayerJob.JobLevels[type]];
+                var ratio = (float)dbPlayerJob.JobXP[type] / (float)XP[dbPlayerJob.JobLevels[type]];
                 jobProgresses.Add(Math.Clamp(ratio, 0f, 1f));
             }
